feat: Add functionality to create a new project
This commit adds the ability to create a new project by initializing it with npm and copying all files and subdirectories from the root directory to the specified destination. Excluded directories, such as "node_modules" and "bin", are not copied. The user is prompted to provide a directory path for the new project. If successful, a success message is displayed; otherwise, an error message is shown.