feat: Add Cloudflare pages deployment
This commit adds the ability to deploy a package to Cloudflare pages. The `deployToCloudflare` function now takes two parameters: `packageName` and `destination`. It sets up a Cloudflare pages project using the provided package name and deploys it to the specified destination directory. If connected to GitHub, continuous deployment is set up; otherwise, a manual deploy is performed.