Refactor Navbar component to dynamically update open state on window resize
The Navbar component has been refactored to include a new `x-data` attribute that initializes an object with properties for `open` and `windowWidth`. The `init()` function is called on component initialization and sets the initial value of `open` based on the window width. Additionally, a `$watch` method is used to update the value of `open` whenever the window width changes. A resize event listener has also been added to update the value of `windowWidth` when the window is resized. This ensures that the Navbar's open state is dynamically updated based on the window size.